+++ plugins/colors/MetricMapping.cpp
@@ -268,7 +268,7 @@ public:
     composite->addGlEntity(label1,"min");
     composite->addGlEntity(label2,"max");
 
-    mainDataSet.set<int>("caption",(int)composite);
+    mainDataSet.set<int>("caption",(long)composite);
     entityDataSet.set<DataSet>("Foreground",mainDataSet);
     dataSet->set<DataSet>("entities",entityDataSet);
